Small fix to AsyncRemoteStrategy
Found a bug while working on getting the goblin test suite to run
against DSE Graph.

Previously, when creating a remote graph with a `driver.Client` instead
of a `driver.Connection`, traversals were always submitted with the eval
processor, even when using GraphSON2.

This is because AsyncRemoteStrategy determined which op processor to used
based on the check `isinstance(self.remote_connection.message_serializer,
		  GraphSON2MessageSerializer)`. On client instances,
`client.message_serializer` is a class, not an instance, so that check
always returned false.

This makes it possible to use the traversal op processor & side effects
API with custom aliases (since aliases can be set on clients but not
			 connections)