removed connection kwargs from ConnectionPool.acquire
I realized these connection creation kwargs on ConnectionPool.acquire
were a potential issue, because while they would be used if a new
connection was created, if a connection was available it would be
acquired despite possibly having a different configuration entirely

Also, if different kwargs were passed in successive calls to acquire you could
end up with a pool of connections with different configurations and no
way of knowing which configuration you'd end up with after calling
acquire.

None of our code was using these.